// src/modules/payments/interfaces/paymob-payment-key-response.interface.ts
export interface PaymobPaymentKeyResponse {
  token: string;
}

export interface BillingData {
  apartment: string;
  email: string;
  floor: string;
  first_name: string;
  street: string;
  building: string;
  phone_number: string;
  shipping_method: string;
  postal_code: string;
  city: string;
  country: string;
  last_name: string;
  state: string;
}